Minor fixes

+++ src/symmetry.h
@@ -104,7 +104,7 @@ namespace ALM_NS
                   const int verbosity,
                   Timer *timer);
 
-        size_t nsym, ntran, nat_prim;
+        unsigned int nsym, ntran, nat_prim;
         std::vector<int> symnum_tran;
 
         double tolerance;
